Kleine Änderungen für einen pull request

+++ block_tracker_setup.sh
@@ -8,12 +8,11 @@ on_error() {
 [ $UID -eq 0 ] || on_error "Du musst root sein"
 
 if [ $# -gt 0 ] && [ $1 == "--uninstall" ]; then
-	echo "Stelle /etc/hosts wider her und "
-	echo "lösche /etc/hosts.d"
+	echo "Stelle /etc/hosts wieder her und lösche /etc/hosts.d"
 	echo "Fortsetzen? [j/N]"
 	read -sn 1 response
-	response=$(echo "${response}" | tr '[:upper:]' '[:lower:]')
-	[ ${response} == j ] || on_error "Abgebrochen!"
+	response=${response,,}
+	[ "${response}" != "j" ] && exit 1	# just quit
 	echo "Deinstalliere block-tracker..." 
 	cp /etc/hosts.d/00-hosts /etc/hosts
 	rm -r /etc/hosts.d/
